Rr 1.0

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/telop/MainTelop.kt
@@ -93,7 +93,7 @@ class MainTelop: LinearOpMode()
 				drive.setLocalizer(localizer);
 				localizer.update();
 				imuHeading = imu.position.h;
-				localHeading = localizer.poseEstimate.heading;
+				localHeading = localizer.poseEstimate.heading.toDouble();
 			}
 
 			//Horizontal Slides
